+++ spirit/extra/project_template/project_name/urls.py
@@ -17,15 +17,9 @@ admin.site.login = login_required(admin.site.login)
 urlpatterns = [
     url(r'^', include(spirit.urls)),
 
-    # Examples:
-    # url(r'^$', 'example.views.home', name='home'),
-    # url(r'^example/', include('example.foo.urls')),
-
-    # Uncomment the admin/doc line below to enable admin documentation:
-    # url(r'^admin/doc/', include('django.contrib.admindocs.urls')),
-
-    # Uncomment the next line to enable the admin:
-    url(r'^admin/', include(admin.site.urls)),
+    # This is the default django admin
+    # it's not needed to use Spirit
+    url(r'^admin/', admin.site.urls),
 ]
 
 if settings.DEBUG: